Split builtin_helpers into argument/conversion/collection modules
Decompose builtin_helpers.rs (24.43) into:
- argument.rs: arg count/type guards (7.36)
- conversion.rs: Value to primitive (4.50)
- collection.rs: list element collectors + fold (14.85)

All evaluator files now under 22 complexity.
